Abstract function for staged API
